+++ src/test/java/br/edu/ifg/figuras/RetanguloTest.java
@@ -37,14 +37,14 @@ public class RetanguloTest {
     @Test
     public void deveCalcularAreaCorretamente(){
         Retangulo retangulo = new Retangulo(4L,8L);
-        assertEquals(32L,(long) retangulo.getArea());
+        assertEquals(32L,(double) retangulo.getArea());
         
     }
     
     @Test
     public void deveCalcularPerimetroCorretamente(){
         Retangulo retangulo = new Retangulo(4L,8L);
-        assertEquals(21L,(long) retangulo.getPerimetro());   
+        assertEquals(21L,(double) retangulo.getPerimetro());   
     }
 
 }
